# apps/api-fastapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nha Tro Manager API",
    "description": "H\u1ec7 th\u1ed1ng RESTful API Qu\u1ea3n l\u00fd Nh\u00e0 tr\u1ecd theo ki\u1ebfn tr\u00fac T\u00e1ch bi\u1ec7t FastAPI & React SPA",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/payments": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Get Payments",
        "operationId": "get_payments_api_v1_payments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "invoiceId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Invoiceid"
            }
          },
          {
            "name": "buildingId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Buildingid"
            }
          },
          {
            "name": "method",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PaymentMethod"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Method"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PaymentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fromdate"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Todate"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Pagesize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_PaginatedData_PaymentResponseSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/manual": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Create Manual Payment",
        "operationId": "create_manual_payment_api_v1_payments_manual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentManualCreateSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_PaymentResponseSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/payments/{payment_id}": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Get Payment Detail",
        "operationId": "get_payment_detail_api_v1_payments__payment_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Payment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_PaymentResponseSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/{payment_id}/cancel": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Cancel Payment",
        "operationId": "cancel_payment_api_v1_payments__payment_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Payment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCancelSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_PaymentResponseSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/{payment_id}/refund": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Refund Payment",
        "operationId": "refund_payment_api_v1_payments__payment_id__refund_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Payment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRefundSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_PaymentResponseSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tenant/invoices/{invoice_id}/payments": {
      "get": {
        "tags": [
          "Tenant Payments"
        ],
        "summary": "Get Tenant Invoice Payments",
        "operationId": "get_tenant_invoice_payments_api_v1_tenant_invoices__invoice_id__payments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Invoice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_List_PaymentResponseSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenant/payments": {
      "get": {
        "tags": [
          "Tenant Payments"
        ],
        "summary": "Get Tenant All Payments",
        "operationId": "get_tenant_all_payments_api_v1_tenant_payments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_List_PaymentResponseSchema__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_List_PaymentResponseSchema__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PaymentResponseSchema"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid"
          }
        },
        "type": "object",
        "title": "APIResponse[List[PaymentResponseSchema]]"
      },
...
      "APIResponse_PaginatedData_PaymentResponseSchema__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginatedData_PaymentResponseSchema_"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid"
          }
        },
        "type": "object",
        "title": "APIResponse[PaginatedData[PaymentResponseSchema]]"
      },
...
      "APIResponse_PaymentResponseSchema_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaymentResponseSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid"
          }
        },
        "type": "object",
        "title": "APIResponse[PaymentResponseSchema]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedData_PaymentResponseSchema_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PaymentResponseSchema"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "totalPages": {
            "type": "integer",
            "title": "Totalpages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "title": "PaginatedData[PaymentResponseSchema]"
      },
...
      "PaymentCancelSchema": {
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason",
            "description": "L\u00fd do h\u1ee7y giao d\u1ecbch thanh to\u00e1n"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "PaymentCancelSchema"
      },
      "PaymentManualCreateSchema": {
        "properties": {
          "invoiceId": {
            "type": "string",
            "title": "Invoiceid",
            "description": "ID H\u00f3a \u0111\u01a1n thanh to\u00e1n"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Amount",
            "description": "S\u1ed1 ti\u1ec1n thanh to\u00e1n (> 0 VN\u0110)"
          },
          "method": {
            "$ref": "#/components/schemas/PaymentMethod",
            "description": "Ph\u01b0\u01a1ng th\u1ee9c thanh to\u00e1n (CASH, BANK_TRANSFER, OTHER)"
          },
          "paidAt": {
            "type": "string",
            "format": "date-time",
            "title": "Paidat",
            "description": "Th\u1eddi gian th\u1ef1c hi\u1ec7n thanh to\u00e1n"
          },
          "transactionRef": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transactionref",
            "description": "M\u00e3 giao d\u1ecbch tham chi\u1ebfu"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "invoiceId",
          "amount",
          "method"
        ],
        "title": "PaymentManualCreateSchema"
      },
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "VIETQR",
          "BANK_TRANSFER",
          "CASH",
          "BANK_WEBHOOK",
          "OTHER"
        ],
        "title": "PaymentMethod"
      },
      "PaymentRefundSchema": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Amount",
            "description": "S\u1ed1 ti\u1ec1n ho\u00e0n tr\u1ea3 (> 0 VN\u0110)"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason",
            "description": "L\u00fd do ho\u00e0n ti\u1ec1n"
          },
          "refundedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Refundedat"
          }
        },
        "type": "object",
        "required": [
          "amount",
          "reason"
        ],
        "title": "PaymentRefundSchema"
      },
      "PaymentResponseSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "invoiceId": {
            "type": "string",
            "title": "Invoiceid"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "method": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "paidAt": {
            "type": "string",
            "format": "date-time",
            "title": "Paidat"
          },
          "transactionRef": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transactionref"
          },
          "maskedTransactionRef": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maskedtransactionref"
          },
          "refundAmount": {
            "type": "integer",
            "title": "Refundamount"
          },
          "netAmount": {
            "type": "integer",
            "title": "Netamount"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "cancelledAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelledat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "invoiceId",
          "amount",
          "method",
          "status",
          "paidAt",
          "netAmount",
          "createdAt"
        ],
        "title": "PaymentResponseSchema"
      },
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "PENDING_REVIEW",
          "CONFIRMED",
          "REJECTED",
          "CANCELLED",
          "REFUNDED",
          "PARTIALLY_REFUNDED"
        ],
        "title": "PaymentStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}